mode HomeViewModel to core module
