+++ android/src/main/java/com/mindera/people/android/MinderaPeopleApp.kt
@@ -1,10 +1,7 @@
 package com.mindera.people.android
 
 import android.app.Application
-import com.mindera.people.android.ui.home.HomeViewModel
 import com.mindera.people.startSdk
-import org.koin.androidx.viewmodel.dsl.viewModelOf
-import org.koin.dsl.module
 
 class MinderaPeopleApp : Application() {
 
@@ -14,9 +11,6 @@ class MinderaPeopleApp : Application() {
     }
 
     private fun setupKoin() {
-        startSdk(app = this@MinderaPeopleApp,
-                 appModule = module {
-                     viewModelOf(::HomeViewModel)
-                 })
+        startSdk(app = this@MinderaPeopleApp)
     }
 }
