+++ buildSrc/src/main/kotlin/scripts/kmm_android.gradle.kts
@@ -49,6 +49,7 @@ kotlin {
                 implementation(Dependencies.ktorClientOkHttp)
                 implementation(Dependencies.androidViewModel)
                 implementation(Dependencies.androidViewModelSavedState)
+                implementation(Dependencies.koinAndroid)
             }
         }
         val androidTest by getting {
