+++ core/src/commonMain/kotlin/com/mindera/people/KoinSetup.kt
@@ -1,6 +1,5 @@
 package com.mindera.people
 
-import com.mindera.people.auth.authModule
 import com.mindera.people.data.dataModule
 import com.mindera.people.user.userModule
 import org.koin.core.context.startKoin
@@ -11,12 +10,7 @@ import org.koin.dsl.KoinAppDeclaration
 
 fun initKoin(appDeclaration: KoinAppDeclaration = {}) = startKoin {
     appDeclaration()
-    modules(
-        platformModule,
-        dataModule,
-        userModule,
-        authModule
-    )
+    modules(platformModule, dataModule, userModule)
 }
 
 inline fun <reified T> Scope.getWith(vararg params: Any?): T =
