--- core/src/commonMain/kotlin/com/mindera/people/auth/AuthModule.kt
@@ -1,8 +1,0 @@
-package com.mindera.people.auth
-
-import org.koin.core.module.dsl.factoryOf
-import org.koin.dsl.module
-
-val authModule = module {
-    factoryOf(::AuthViewModel)
-}
